[ALPS10473608] EMI: Fix EMI debug dump handling for TFA readiness
[Description]
Updated EMI debug dump logic to correctly handle cases where TFA is
not ready, ensuring proper processing when smc_res.a0 returns -1.
add in v2 & v1

[Test]
Local test

MTK-Commit-Id: 69db59e645a9e86706d52b17c9f34c54b5e62174


CR-Id: ALPS10473608
Feature: emi

+++ drivers/memory/mediatek/emi_legacy/emi_legacy_v1/emicen.c
@@ -103,7 +103,7 @@ void mtk_emidbg_dump(void)
 
 	arm_smccc_smc(MTK_SIP_EMIMPU_CONTROL, MTK_EMIDBG_DUMP,
 		0, 0, 0, 0, 0, 0, &smc_res);
-	while (smc_res.a0 > 0) {
+	while (((long)smc_res.a0) > 0) {
 		arm_smccc_smc(MTK_SIP_EMIMPU_CONTROL, MTK_EMIDBG_MSG,
 		0, 0, 0, 0, 0, 0, &smc_res);
